+++ CHANGELOG.rst
@@ -25,6 +25,7 @@ Common
      intercept API calls
    * client ids and "get_neighbor_clients" API method (requires EVA JS Framework
      0.3.24+)
+   * notifier frame counter and online info with "notifier list" CLI command
 
 UC
 --